+++ src/Logger.php
@@ -13,17 +13,14 @@ use Seablast\Logger\LoggerTime;
 class Logger extends AbstractLogger implements LoggerInterface
 {
     // Define constants for configuration keys
-    // phpcs:disable PSR12.Properties.ConstantVisibility.NotFound
-    // todo remove phpcs exception when PHP/5 support removed
-    const CONF_ERROR_LOG_MESSAGE_TYPE = 'error_log_message_type';
-    const CONF_LOGGING_FILE = 'logging_file';
-    const CONF_LOGGING_LEVEL = 'logging_level';
-    const CONF_LOGGING_LEVEL_NAME = 'logging_level_name';
-    const CONF_LOGGING_LEVEL_PAGE_SPEED = 'logging_level_page_speed';
-    const CONF_LOG_MONTHLY_ROTATION = 'log_monthly_rotation';
-    const CONF_LOG_PROFILING_STEP = 'log_profiling_step';
-    const CONF_MAIL_FOR_ADMIN_ENABLED = 'mail_for_admin_enabled';
-    // phpcs:enable
+    public const CONF_ERROR_LOG_MESSAGE_TYPE = 'error_log_message_type';
+    public const CONF_LOGGING_FILE = 'logging_file';
+    public const CONF_LOGGING_LEVEL = 'logging_level';
+    public const CONF_LOGGING_LEVEL_NAME = 'logging_level_name';
+    public const CONF_LOGGING_LEVEL_PAGE_SPEED = 'logging_level_page_speed';
+    public const CONF_LOG_MONTHLY_ROTATION = 'log_monthly_rotation';
+    public const CONF_LOG_PROFILING_STEP = 'log_profiling_step';
+    public const CONF_MAIL_FOR_ADMIN_ENABLED = 'mail_for_admin_enabled';
 
     /** @var array<mixed> int,string,bool,array */
     protected $conf = array();
@@ -91,7 +88,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param int $newLevel
      * @return void
      */
-    public function logAtLeastToLevel(int $newLevel)
+    public function logAtLeastToLevel(int $newLevel): void
     {
         if (!is_int($newLevel)) {
             throw new \Psr\Log\InvalidArgumentException('The variable $newLevel is not an integer.');
@@ -102,7 +99,7 @@ class Logger extends AbstractLogger implements LoggerInterface
     /**
      * @return float
      */
-    public function getLastRunningTime()
+    public function getLastRunningTime(): float
     {
         return $this->runningTime;
     }
@@ -113,7 +110,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param int|string $user
      * @return void
      */
-    public function setUser($user)
+    public function setUser($user): void
     {
         $this->user = (string) $user;
     }
@@ -125,9 +122,8 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function emergency($message, array $context = array())
+    public function emergency($message, array $context = array()): void
     {
-        //TODO when adding PHP/8 support, add :void to the inherited methods and remove PHP/5 support
         $this->log(0, $message, $context);
     }
 
@@ -141,7 +137,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function alert($message, array $context = array())
+    public function alert($message, array $context = array()): void
     {
         $this->log(1, $message, $context);
     }
@@ -155,7 +151,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function critical($message, array $context = array())
+    public function critical($message, array $context = array()): void
     {
         $this->log(1, $message, $context);
     }
@@ -168,7 +164,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function error($message, array $context = array())
+    public function error($message, array $context = array()): void
     {
         $this->log(2, $message, $context);
     }
@@ -183,7 +179,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function warning($message, array $context = array())
+    public function warning($message, array $context = array()): void
     {
         $this->log(3, $message, $context);
     }
@@ -195,7 +191,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function notice($message, array $context = array())
+    public function notice($message, array $context = array()): void
     {
         $this->log(4, $message, $context);
     }
@@ -209,7 +205,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function info($message, array $context = array())
+    public function info($message, array $context = array()): void
     {
         $this->log(4, $message, $context);
     }
@@ -221,7 +217,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      * @param array<int> $context
      * @return void
      */
-    public function debug($message, array $context = array())
+    public function debug($message, array $context = array()): void
     {
         $this->log(5, $message, $context);
     }
@@ -265,7 +261,7 @@ class Logger extends AbstractLogger implements LoggerInterface
      *  1001 Establish correct error_number
      *
      */
-    public function log($level, $message, array $context = array())
+    public function log($level, $message, array $context = array()): void
     {
         //TODO add variable $line - it should always be called as basename(__FILE__)."#".__LINE__ ,
         //so it's clear which line of the source code triggered the call